requirements.txt
Django==5.1.1
djangorestframework==3.15.2
numpy==2.1.2
pandas==2.2.3
scikit_learn==1.5.2
